Remove other web elements extensions, leaving wanted
